pyproject.toml
[tool.poetry]
name = "systems"
version = "0.1.0"
description = "Systems code for Relectric car team."
authors = ["Relectric Car Team <info@teamrelectric.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyzmq = "^22.0.3"
pyserial = "^3.5"
attrs = "^20.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-xdist = "^2.2.1"
pre-commit = "^2.11.0"
pydocstyle = "^5.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 4
line_length = 79
include_trailing_comma = true